requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-compressor==1.4
djangorestframework==2.4.4
drf-nested-routers==0.9.0
gunicorn==19.1.1
psycopg2==2.5.4
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
